+++ src/main/java/com/Interceptor/LoginInterceptor.java
@@ -3,6 +3,7 @@ package com.Interceptor;
 import com.Utils.Response;
 import com.alibaba.fastjson.JSONObject;
 import org.apache.tomcat.util.net.NioEndpoint;
+import org.springframework.web.method.HandlerMethod;
 import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
@@ -20,7 +21,7 @@ public class LoginInterceptor implements HandlerInterceptor {
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
         HttpSession session = request.getSession();
-        Integer userId = (Integer) session.getAttribute("userId");
+        Long userId = (Long) session.getAttribute("userId");
         if (userId == null) {
             Object jsonObject = JSONObject.toJSON(new Response().notVisit(40001, "需要登录"));
             response.setCharacterEncoding("UTF-8");
