+++ src/main/java/com/Service/GroupService.java
@@ -6,9 +6,9 @@ import com.Entity.User;
 import java.util.List;
 
 public interface GroupService {
-    int deleteOne(int groupid, int deleteid);
-    int addOne(int groupid,int addid);
-    int deleteGroup(int groupid,int ownerid);
-    int changeOwner(int groupid,int userid);
-    Groupchat buildGroup(int userid,String groupname);
+    int deleteOne(int groupid, long deleteid);
+    int addOne(int groupid,long addid);
+    int deleteGroup(int groupid,long ownerid);
+    int changeOwner(int groupid,long userid);
+    Groupchat buildGroup(long userid,String groupname);
 }
