+++ src/main/java/com/Service/RedPacketByRedisService.java
@@ -4,10 +4,10 @@ import com.Entity.PubRedPacket;
 import com.Entity.UserRedPacket;
 
 public interface RedPacketByRedisService {
-    PubRedPacket publishRedPacket(int userId, double money, int redPacketType, int groupId, int size);
+    PubRedPacket publishRedPacket(long userId, double money, int redPacketType, int groupId, int size);
 
-    UserRedPacket getRedPacket(long redPacketId, int userId);
+    UserRedPacket getRedPacket(long redPacketId, long userId);
 
-    void persistToSql(long redPacketId,int userid);
+    void persistToSql(long redPacketId,long userid);
 
 }
